refactor(editor): reusable editor component
