+++ apps/web/package.json
@@ -24,6 +24,7 @@
     "@buildit/env": "workspace:^",
     "@buildit/ui": "workspace:^",
     "@buildit/utils": "workspace:^",
+    "@radix-ui/react-visually-hidden": "^1.1.0",
     "@sentry/nextjs": "^8.18.0",
     "@sindresorhus/slugify": "^2.2.1",
     "@t3-oss/env-nextjs": "^0.10.1",
