+++ apps/web/src/components/ui/modal.tsx
@@ -3,12 +3,14 @@
 import React from 'react'
 
 import * as Dialog from '@radix-ui/react-dialog'
+import * as VisuallyHidden from '@radix-ui/react-visually-hidden'
 
 import { cn } from '@buildit/ui/cn'
 
 import { Icons } from '@/components/ui/icons'
 
 export const ModalTrigger = Dialog.Trigger
+export const ModalTitle = Dialog.Title
 
 export const Modal = ({
   children,
@@ -38,9 +40,10 @@ export const ModalContent = ({
       <Dialog.Overlay className='data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 fixed inset-0 z-50 bg-black/10 data-[state=closed]:animate-out data-[state=open]:animate-in' />
       <Dialog.Content
         className={cn(
-          'data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[state=closed]:slide-out-to-left-1/2 data-[state=closed]:slide-out-to-top-[10%] data-[state=open]:slide-in-from-left-1/2 data-[state=open]:slide-in-from-top-[10%] fixed top-[20%] left-[50%] z-50 max-h-[650px] w-full max-w-2xl translate-x-[-50%] translate-y-[-10%] space-y-2 border border-slate-200 bg-white p-4 shadow-lg duration-200 data-[state=closed]:animate-out data-[state=open]:animate-in sm:rounded-lg dark:border-slate-800 dark:bg-slate-950',
+          'data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[state=closed]:slide-out-to-left-1/2 data-[state=closed]:slide-out-to-top-[10%] data-[state=open]:slide-in-from-left-1/2 data-[state=open]:slide-in-from-top-[10%] fixed top-[20%] left-[50%] z-50 max-h-[650px] w-full max-w-2xl translate-x-[-50%] translate-y-[-10%] border border-slate-200 bg-white p-4 flex flex-col gap-y-2 shadow-lg duration-200 data-[state=closed]:animate-out data-[state=open]:animate-in sm:rounded-lg dark:border-slate-800 dark:bg-slate-950',
           className,
         )}
+        aria-describedby={undefined}
       >
         {children}
       </Dialog.Content>
@@ -88,3 +91,22 @@ export const ModalHeader = ({
     </nav>
   )
 }
+
+export const ModalFooter = ({
+  className,
+  ...props
+}: React.HTMLAttributes<HTMLDivElement>) => (
+  <div
+    className={cn(
+      'flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2',
+      className,
+    )}
+    {...props}
+  />
+)
+ModalFooter.displayName = 'ModalFooter'
+
+export const VisuallyHide = ({ children }: { children: React.ReactNode }) => (
+  <VisuallyHidden.Root asChild>{children}</VisuallyHidden.Root>
+)
+VisuallyHide.displayName = 'hidden'
