+++ packages/editor/src/components/editor/index.tsx
@@ -1,82 +1,23 @@
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 'use client'
 
 import { useRef } from 'react'
 
 import type { Value } from '@udecode/plate-common'
 
-import { cn, withProps } from '@udecode/cn'
-import { AutoformatPlugin } from '@udecode/plate-autoformat/react'
-import {
-  BoldPlugin,
-  CodePlugin,
-  ItalicPlugin,
-  StrikethroughPlugin,
-  SubscriptPlugin,
-  SuperscriptPlugin,
-  UnderlinePlugin,
-} from '@udecode/plate-basic-marks/react'
-import { BlockquotePlugin } from '@udecode/plate-block-quote/react'
-import { ExitBreakPlugin, SoftBreakPlugin } from '@udecode/plate-break/react'
-import {
-  isCodeBlockEmpty,
-  isSelectionAtCodeBlockStart,
-  unwrapCodeBlock,
-} from '@udecode/plate-code-block'
-import {
-  CodeBlockPlugin,
-  CodeLinePlugin,
-  CodeSyntaxPlugin,
-} from '@udecode/plate-code-block/react'
-import {
-  isBlockAboveEmpty,
-  isSelectionAtBlockStart,
-  someNode,
-} from '@udecode/plate-common'
-import {
-  createPlateEditor,
-  ParagraphPlugin,
-  Plate,
-  PlateLeaf,
-} from '@udecode/plate-common/react'
-import { HEADING_KEYS, HEADING_LEVELS } from '@udecode/plate-heading'
-import { HeadingPlugin } from '@udecode/plate-heading/react'
-import { HighlightPlugin } from '@udecode/plate-highlight/react'
-import { HorizontalRulePlugin } from '@udecode/plate-horizontal-rule/react'
-import { IndentListPlugin } from '@udecode/plate-indent-list/react'
-import { IndentPlugin } from '@udecode/plate-indent/react'
-import { TodoListPlugin } from '@udecode/plate-list/react'
-import { MentionInputPlugin, MentionPlugin } from '@udecode/plate-mention/react'
-import { NodeIdPlugin } from '@udecode/plate-node-id'
-import { ResetNodePlugin } from '@udecode/plate-reset-node/react'
-import {
-  SlashInputPlugin,
-  SlashPlugin,
-} from '@udecode/plate-slash-command/react'
-import { TabbablePlugin } from '@udecode/plate-tabbable/react'
-import Prism from 'prismjs'
+import { cn } from '@udecode/cn'
+import { Plate } from '@udecode/plate-common/react'
 
-import { autoformatRules } from '../../lib/autoformat-rules'
-import { BlockquoteElement } from '../plate-ui/blockquote-element'
-import { CodeBlockElement } from '../plate-ui/code-block-element'
-import { CodeLeaf } from '../plate-ui/code-leaf'
-import { CodeLineElement } from '../plate-ui/code-line-element'
-import { CodeSyntaxLeaf } from '../plate-ui/code-syntax-leaf'
+import { useMyEditor } from '../../lib/use-my-editor'
 import { Editor as PlateEditor } from '../plate-ui/editor'
 import { FloatingToolbar } from '../plate-ui/floating-toolbar'
 import { FloatingToolbarButtons } from '../plate-ui/floating-toolbar-buttons'
-import { HeadingElement } from '../plate-ui/heading-element'
-import { HrElement } from '../plate-ui/hr-element'
-import { MentionElement } from '../plate-ui/mention-element'
-import { MentionInputElement } from '../plate-ui/mention-input-element'
-import { ParagraphElement } from '../plate-ui/paragraph-element'
-import { withPlaceholders } from '../plate-ui/placeholder'
-import { SlashInputElement } from '../plate-ui/slash-input-element'
-import { TodoListElement } from '../plate-ui/todo-list-element'
 
-interface EditorProps {
-  onBlur: () => void
+export interface EditorProps {
   content: Value
+  onBlur: () => void
+  onChange: (value: Value) => void
+  autoFocus?: boolean
+  readOnly?: boolean
 }
 
 /**
@@ -84,9 +25,18 @@ interface EditorProps {
  * @param props The props for the editor component.
  * @param props.content The initial content of the editor.
  * @param props.onBlur The callback function to call when the content changes.
+ * @param props.onChange
+ * @param props.autoFocus
+ * @param props.readOnly
  * @returns The editor component.
  */
-export default function Editor({ content, onBlur }: EditorProps) {
+export default function Editor({
+  content,
+  onBlur,
+  onChange,
+  autoFocus = false,
+  readOnly = false,
+}: EditorProps) {
   const containerRef = useRef(null)
   const editor = useMyEditor(content)
 
@@ -94,7 +44,7 @@ export default function Editor({ content, onBlur }: EditorProps) {
     <Plate
       editor={editor}
       onChange={({ value }) => {
-        localStorage.setItem('editorContent', JSON.stringify(value))
+        onChange(value)
       }}
     >
       <div
@@ -107,10 +57,12 @@ export default function Editor({ content, onBlur }: EditorProps) {
       >
         <PlateEditor
           focusRing={false}
-          autoFocus
-          onBlur={() => {
-            onBlur()
-          }}
+          autoFocus={autoFocus}
+          onBlur={onBlur}
+          variant={'ghost'}
+          readOnly={readOnly}
+          className='px-0'
+          placeholder=''
         />
         <FloatingToolbar>
           <FloatingToolbarButtons />
@@ -118,185 +70,4 @@ export default function Editor({ content, onBlur }: EditorProps) {
       </div>
     </Plate>
   )
-}
-
-export const useMyEditor = (content: Value) => {
-  const editor = createPlateEditor({
-    plugins: [
-      // Nodes
-      HeadingPlugin,
-      BlockquotePlugin,
-      CodeBlockPlugin.configure({
-        options: {
-          prism: Prism,
-        },
-      }),
-      CodeLinePlugin,
-      CodeSyntaxPlugin,
-      HorizontalRulePlugin,
-      TodoListPlugin,
-      SlashPlugin,
-      MentionPlugin,
-      MentionInputPlugin,
-
-      // Marks
-      BoldPlugin,
-      ItalicPlugin,
-      UnderlinePlugin,
-      StrikethroughPlugin,
-      CodePlugin,
-      SubscriptPlugin,
-      SuperscriptPlugin,
-      HighlightPlugin,
-
-      // Block Style
-
-      IndentPlugin.configure({
-        inject: {
-          targetPlugins: [
-            ParagraphPlugin.key,
-            BlockquotePlugin.key,
-            CodeBlockPlugin.key,
-            ...HEADING_LEVELS,
-          ],
-        },
-      }),
-      IndentListPlugin.extend({
-        inject: {
-          targetPlugins: [
-            ParagraphPlugin.key,
-            BlockquotePlugin.key,
-            CodeBlockPlugin.key,
-            ...HEADING_LEVELS,
-          ],
-        },
-      }),
-
-      // Functionality
-      AutoformatPlugin.configure({
-        options: {
-          rules: autoformatRules,
-          enableUndoOnDelete: true,
-        },
-      }),
-      ExitBreakPlugin.configure({
-        options: {
-          rules: [
-            {
-              hotkey: 'mod+enter',
-            },
-            {
-              hotkey: 'mod+shift+enter',
-              before: true,
-            },
-            {
-              hotkey: 'enter',
-              query: {
-                start: true,
-                end: true,
-                allow: HEADING_LEVELS,
-              },
-              relative: true,
-              level: 1,
-            },
-          ],
-        },
-      }),
-      NodeIdPlugin,
-      ResetNodePlugin.configure({
-        options: {
-          rules: [
-            {
-              types: [BlockquotePlugin.key, TodoListPlugin.key],
-              defaultType: ParagraphPlugin.key,
-              hotkey: 'Enter',
-              predicate: isBlockAboveEmpty,
-            },
-            {
-              types: [BlockquotePlugin.key, TodoListPlugin.key],
-              defaultType: ParagraphPlugin.key,
-              hotkey: 'Backspace',
-              predicate: isSelectionAtBlockStart,
-            },
-            {
-              types: [CodeBlockPlugin.key],
-              defaultType: ParagraphPlugin.key,
-              onReset: unwrapCodeBlock,
-              hotkey: 'Enter',
-              predicate: isCodeBlockEmpty,
-            },
-            {
-              types: [CodeBlockPlugin.key],
-              defaultType: ParagraphPlugin.key,
-              onReset: unwrapCodeBlock,
-              hotkey: 'Backspace',
-              predicate: isSelectionAtCodeBlockStart,
-            },
-          ],
-        },
-      }),
-      SoftBreakPlugin.configure({
-        options: {
-          rules: [
-            { hotkey: 'shift+enter' },
-            {
-              hotkey: 'enter',
-              query: {
-                allow: [CodeBlockPlugin.key, BlockquotePlugin.key],
-              },
-            },
-          ],
-        },
-      }),
-      //! Fix: Tabbable Plugin is not working - when pressing tab, the focus is not moving to the next element instead it is focusing the browser's window
-      TabbablePlugin.configure(({ editor }) => ({
-        options: {
-          query: () => {
-            if (isSelectionAtBlockStart(editor)) return false
-
-            return !someNode(editor, {
-              match: (n) => {
-                return !!(
-                  n.type &&
-                  ([TodoListPlugin.key, CodeBlockPlugin.key].includes(
-                    n.type as string,
-                  ) ||
-                    n[IndentListPlugin.key])
-                )
-              },
-            })
-          },
-        },
-      })),
-    ],
-    override: {
-      components: withPlaceholders({
-        [BlockquotePlugin.key]: BlockquoteElement,
-        [CodeBlockPlugin.key]: CodeBlockElement,
-        [CodeLinePlugin.key]: CodeLineElement,
-        [CodeSyntaxPlugin.key]: CodeSyntaxLeaf,
-        [HorizontalRulePlugin.key]: HrElement,
-        [HEADING_KEYS.h1]: withProps(HeadingElement, { variant: 'h1' }),
-        [HEADING_KEYS.h2]: withProps(HeadingElement, { variant: 'h2' }),
-        [HEADING_KEYS.h3]: withProps(HeadingElement, { variant: 'h3' }),
-        [HEADING_KEYS.h4]: withProps(HeadingElement, { variant: 'h4' }),
-        [HEADING_KEYS.h5]: withProps(HeadingElement, { variant: 'h5' }),
-        [HEADING_KEYS.h6]: withProps(HeadingElement, { variant: 'h6' }),
-        [SlashInputPlugin.key]: SlashInputElement,
-        [MentionPlugin.key]: MentionElement,
-        [MentionInputPlugin.key]: MentionInputElement,
-        [ParagraphPlugin.key]: ParagraphElement,
-        [TodoListPlugin.key]: TodoListElement,
-        [BoldPlugin.key]: withProps(PlateLeaf, { as: 'strong' }),
-        [CodePlugin.key]: CodeLeaf,
-        [ItalicPlugin.key]: withProps(PlateLeaf, { as: 'em' }),
-        [StrikethroughPlugin.key]: withProps(PlateLeaf, { as: 's' }),
-        [SubscriptPlugin.key]: withProps(PlateLeaf, { as: 'sub' }),
-        [SuperscriptPlugin.key]: withProps(PlateLeaf, { as: 'sup' }),
-        [UnderlinePlugin.key]: withProps(PlateLeaf, { as: 'u' }),
-      }),
-    },
-    value: content,
-  })
-  return editor
 }
