+++ packages/utils/src/validations/index.ts
@@ -1,4 +1,5 @@
 export * from './auth'
 export * from './getting-started'
+export * from './issue'
 export * from './projects'
 export * from './team'
