+++ packages/utils/src/validations/issue/index.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod'
+
+export const CreateIssueSchema = z.object({
+  title: z.string(),
+  description: z.any().nullable(),
+})
